backend/pyproject.toml
[tool.poetry]
name = "promptly-backend"
version = "1.0.0"
description = "FastAPI backend for Promptly - Interactive AI Prompting Platform"
authors = ["Promptly Team"]
readme = "../README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
motor = "^3.3.2"
pydantic = {extras = ["email"], version = "^2.5.0"}
fastapi-users = {extras = ["mongodb", "jwt", "oauth"], version = "^12.1.2"}
argon2-cffi = "^23.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx-oauth = "^0.16.1"
httpx = "^0.28.1"
slowapi = "^0.1.9"
redis = "^5.0.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
ruff = "^0.1.6"
mypy = "^1.7.1"

[tool.poetry.scripts]
api = "uvicorn backend.main:app --host 0.0.0.0 --port 8000 --reload"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by ruff format
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "motor.*",
    "slowapi.*",
]
ignore_missing_imports = true 